requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
async-timeout==4.0.2
attrs==21.4.0
binaryornot==0.4.4
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
clip==1.0
cloudpickle==2.0.0
cookiecutter==1.7.3
cycler==0.11.0
dask==2022.4.1
filelock==3.6.0
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
google-auth==2.6.5
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
Jinja2==3.1.1
jinja2-time==0.2.0
joblib==1.1.0
kaggle==1.5.12
kiwisolver==1.4.2
locket==0.2.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
multidict==6.0.2
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
partd==1.2.0
Pillow==9.1.0
plotly==5.11.0
poyo==0.5.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-slugify==6.1.1
pytorch-lightning==1.6.1
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
tenacity==8.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
threadpoolctl==3.1.0
tokenizers==0.12.1
toolz==0.11.2
torch==1.11.0
torchmetrics==0.8.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.8.0
